+++ src/algorithms/histogram.cpp
@@ -9,13 +9,13 @@ void Histogram::apply8bit(std::shared_ptr<Image> &image)
 
     auto size = image->pixelData()->size();
 
-    for (uint64_t n = 0; n < size; n++) {
+    for (int32_t n = 0; n < size; n++) {
         auto val = static_cast<uint8_t>(image->pixelData()->at(n));
         px.at(val) = px.at(val) + 1;
     }
 
     // 1a. Normalized (divide each occurrence in each bin by number of pixels in the image)
-    for (int i = 0; i < 256; i++) {
+    for (int32_t i = 0; i < 256; i++) {
         auto normalizedFrequency = float(px.at(i)) / float(size);
         px.at(i) = normalizedFrequency;
     }
@@ -35,7 +35,7 @@ bool Histogram::write(const std::string &filePath)
 
     auto size = mHistogramBin.size();
 
-    for (int i = 0; i < size; i++) {
+    for (int32_t i = 0; i < size; i++) {
         ofs << std::to_string(mHistogramBin.at(i)) << std::endl;
     }
